+++ cmp/tab/TabContainer.ts
@@ -4,18 +4,33 @@
  *
  * Copyright © 2025 Extremely Heavy Industries Inc.
  */
+import {TabSwitcherProps} from '@xh/hoist/cmp/tab/Types';
 import {BoxProps, hoistCmp, HoistProps, refreshContextView, uses, XH} from '@xh/hoist/core';
 import {tabContainerImpl as desktopTabContainerImpl} from '@xh/hoist/dynamics/desktop';
 import {tabContainerImpl as mobileTabContainerImpl} from '@xh/hoist/dynamics/mobile';
 import {TabContainerModel} from './TabContainerModel';
 
-export type TabContainerProps = HoistProps<TabContainerModel> & BoxProps;
+export interface TabContainerProps extends HoistProps<TabContainerModel>, BoxProps {
+    /**
+     * Props to apply to child TabContainers, if any. Specify as an object or a function
+     * that returns props based on the tabId and depth of the child container.
+     */
+    childTabContainerProps?:
+        | TabContainerProps
+        | ((ctx: {tabId: string; depth: number}) => TabContainerProps);
+    /**
+     * Indicates whether to include a default switcher docked within this component. Specify as a
+     * boolean or an object containing props for a TabSwitcher component. Set to false to not
+     * include a switcher. Defaults to true.
+     */
+    switcher?: boolean | TabSwitcherProps;
+}
 
 /**
  * Display a set of child Tabs and (optionally) a switcher control.
  *
- * By default this TabContainer will install a TabSwitcher above the Tabs to control the currently
- * displayed Tab. The 'TabContainerModel.switcher' property can be adjusted to place the switcher
+ * By default, this TabContainer will install a TabSwitcher above the Tabs to control the currently
+ * displayed Tab. The 'TabContainerProps.switcher' property can be adjusted to place the switcher
  * control on alternative edges of the container.
  *
  * If `switcher` is set to false then no TabSwitcher will be installed.  This setting