+++ desktop/cmp/tab/impl/TabContainer.ts
@@ -6,20 +6,33 @@
  */
 import {div, hbox, placeholder, vbox} from '@xh/hoist/cmp/layout';
 import {TabContainerModel, TabContainerProps} from '@xh/hoist/cmp/tab';
+import {TabSwitcherProps} from '@xh/hoist/cmp/tab/Types';
 import {getTestId} from '@xh/hoist/utils/js';
 import {getLayoutProps} from '@xh/hoist/utils/react';
-import {isEmpty} from 'lodash';
+import {isEmpty, isObject} from 'lodash';
 import '../Tabs.scss';
 import {tabSwitcher} from '../TabSwitcher';
+import {dynamicTabSwitcher} from '../dynamic/DynamicTabSwitcher';
 import {tab} from './Tab';
 
 /**
  * Desktop implementation of TabContainer.
  * @internal
  */
-export function tabContainerImpl({model, className, testId, ...props}: TabContainerProps) {
-    const layoutProps = getLayoutProps(props),
-        vertical = ['left', 'right'].includes(model.switcher?.orientation),
+export function tabContainerImpl({
+    model,
+    childTabContainerProps,
+    className,
+    testId,
+    ...props
+}: TabContainerProps) {
+    const switcher: TabSwitcherProps = isObject(props.switcher)
+            ? props.switcher
+            : props.switcher === false
+              ? null
+              : {orientation: 'top'},
+        layoutProps = getLayoutProps(props),
+        vertical = ['left', 'right'].includes(switcher?.orientation),
         container = vertical ? hbox : vbox;
 
     // Default flex = 'auto' if no dimensions / flex specified.
@@ -31,21 +44,27 @@ export function tabContainerImpl({model, className, testId, ...props}: TabContai
         ...layoutProps,
         className,
         testId,
-        item: getChildren(model, testId)
+        item: getChildren(model, switcher, testId, childTabContainerProps)
     });
 }
 
-function getChildren(model: TabContainerModel, testId: string) {
+function getChildren(
+    model: TabContainerModel,
+    switcher: TabSwitcherProps,
+    testId: string,
+    childTabContainerProps: TabContainerProps['childTabContainerProps']
+) {
     const {tabs} = model;
     if (isEmpty(tabs)) {
         return div({className: 'xh-tab-wrapper', item: placeholder(model.emptyText)});
     }
 
-    const {activeTabId, switcher} = model,
+    const {activeTabId, dynamicTabSwitcherModel} = model,
         switcherBefore = ['left', 'top'].includes(switcher?.orientation),
         switcherAfter = ['right', 'bottom'].includes(switcher?.orientation),
+        switcherImpl = dynamicTabSwitcherModel ? dynamicTabSwitcher : tabSwitcher,
         switcherCmp = switcher
-            ? tabSwitcher({key: 'switcher', testId: getTestId(testId, 'switcher'), ...switcher})
+            ? switcherImpl({key: 'switcher', testId: getTestId(testId, 'switcher'), ...switcher})
             : null;
 
     return [
@@ -58,7 +77,11 @@ function getChildren(model: TabContainerModel, testId: string) {
                 className: 'xh-tab-wrapper',
                 style,
                 key: tabId,
-                item: tab({model: tabModel, testId: getTestId(testId, tabId)})
+                item: tab({
+                    childTabContainerProps,
+                    model: tabModel,
+                    testId: getTestId(testId, tabId)
+                })
             });
         }),
         switcherAfter ? switcherCmp : null