+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 ## 78.0.0-SNAPSHOT - unreleased
 
+### 🎁 New Features
+
+* Added new `DynamicTabSwitcher` component, a more user-customizable version of `TabSwitcher` that
+  allows for dynamic addition, removal, and drag-and-drop reordering of tabs with the ability to
+  persist "favorited" tab state across sessions. See `TabContainerConfig.dynamicTabSwitcherModel`.
+
+### 💥 Breaking Changes
+* `TabSwitcherProps` has moved to `cmp/tab/Types.ts` but is still exported from `cmp/tab/index.ts`.
+  Some apps may need to update their imports.
+* `TabContainerConfig.switcher` has been moved to `TabContainerProps.switcher`.
+
 ## 77.1.1 - 2025-11-12
 
 ### 🎁 New Features